Moved formatGraph to separate file to have formatAssertion function backwards compatible
