+++ index.js
@@ -1,8 +1,9 @@
-import calculateRoot from './calculate-root.js';
-import assertionFormatting from './assertion-formatting.js';
-import peerId2Hash from './peer-id-2-hash.js';
-import getMerkleProof from './get-merkle-proof.js';
-import keccak256 from './keccak256.js';
 import assertionMetadata from './assertion-metadata.js';
+import calculateRoot from './calculate-root.js';
+import formatAssertion from './assertion-formatting.js';
+import formatGraph from './graph-formatting.js';
+import getMerkleProof from './get-merkle-proof.js';
+import peerId2Hash from './peer-id-2-hash.js';
+import keccak256 from './keccak256.js';
 
-export { calculateRoot, assertionFormatting, peerId2Hash, getMerkleProof, keccak256, assertionMetadata };
+export { calculateRoot, formatAssertion, formatGraph, getMerkleProof, peerId2Hash, keccak256, assertionMetadata };
